refactor: split main.go (1828 lines) into 14 files per command
- main.go: dispatch, command table, top-level help (250 lines)
- help.go: all helpXxx functions
- common.go: flags, connect, popHost
- format.go: all value/tree formatting
- cmd_*.go: one file per CLI command (info, walk, get, set, watch, export, import, discover, diag, list)

No functional changes — pure file reorganization.
